Refactor MPMarketingChannel.
The aim is to make both the MPMarketingChannelController and the MPMarketingChannel class lighter by moving most of the data request feature to the MPMarketingChannelDataController class

MAILPOET-5698
